+++ src/components/Tables/DomainTable/components/HighestBid.tsx
