+++ src/components/Tables/DomainTable/components/NumBids.tsx
